requirements.txt
requests
PyYAML
questionary
python-dotenv 